Split .core package into two packages
