+++ ssltest/core/ratable/CipherSuites.py
@@ -1,13 +1,12 @@
 import logging
-
 from typing import Tuple
 
 from .CipherSuite import CipherSuite
-from .ClientHello import ClientHello
-from .SSLv2 import SSLv2
-from ..exceptions.ConnectionTimeout import ConnectionTimeout
-from ..main.utils import send_data_return_sock, parse_cipher_suite, bytes_to_cipher_suite, protocol_version_conversion, \
-    is_server_hello, get_cipher_suite_protocols, Address
+from ..ClientHello import ClientHello
+from ..SSLv2 import SSLv2
+from ...exceptions.ConnectionTimeout import ConnectionTimeout
+from ...main.utils import send_data_return_sock, parse_cipher_suite, bytes_to_cipher_suite, \
+    protocol_version_conversion, is_server_hello, get_cipher_suite_protocols, Address
 
 log = logging.getLogger(__name__)
