+++ ssltest/core/ratable/ProtocolSupport.py
@@ -3,10 +3,10 @@ import socket
 
 from .PType import PType
 from .Parameters import Parameters
-from .SSLv2 import SSLv2
-from .SSLv3 import SSLv3
-from .connection_utils import create_session, create_ssl_context
-from ..main.utils import Address
+from ..SSLv2 import SSLv2
+from ..SSLv3 import SSLv3
+from ..connection_utils import create_session, create_ssl_context
+from ...main.utils import Address
 
 log = logging.getLogger(__name__)
